+++ apollo-execution-federation/src/commonMain/kotlin/com/apollographql/execution/federation/GraphQLKey.kt
@@ -1,5 +1,0 @@
-package com.apollographql.execution.federation
-
-
-@Target(AnnotationTarget.PROPERTY, AnnotationTarget.FUNCTION)
-annotation class GraphQLKey